+++ packages/path_provider/path_provider_linux/lib/path_provider_linux.dart
@@ -13,7 +13,7 @@ import 'package:path_provider_platform_interface/path_provider_platform_interfac
 /// This class implements the `package:path_provider` functionality for linux
 class PathProviderLinux extends PathProviderPlatform {
   /// Registers this class as the default instance of [PathProviderPlatform]
-  static void register() {
+  static void registerWith() {
     PathProviderPlatform.instance = PathProviderLinux();
   }