+++ packages/path_provider/path_provider_windows/lib/src/path_provider_windows_real.dart
@@ -47,6 +47,11 @@ class VersionInfoQuerier {
 ///
 /// This class implements the `package:path_provider` functionality for Windows.
 class PathProviderWindows extends PathProviderPlatform {
+  /// Registers the Windows implementation.
+  static void registerWith() {
+    // TODO(stuartmorgan): Fill this up.
+  }
+
   /// The object to use for performing VerQueryValue calls.
   @visibleForTesting
   VersionInfoQuerier versionInfoQuerier = VersionInfoQuerier();