+++ packages/path_provider/path_provider_windows/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.0.0-nullsafety.1
+
+* Add `implements` to pubspec.yaml.
+* Add `registerWith()` to the Dart main class.
+
 ## 2.0.0-nullsafety
 
 * Update version to (semi-belatedly) meet 1.0-consistency promise.